Use the smallest underlying type for the enums (performance-enum-size)
- The enums outside FslBase use the smallest underlying type that holds their values, and the
  headers include <cstdint> for it. TextureInfo.hpp forward declares TextureType with the same
  underlying type.
- Enums that are created from file data or command line values (AnimationMode, BitmapFontType,
  ChunkType, DemoScene, InputCommandId, PrimitiveType, RenderMode, VersionId, VertexElementFormat
  and VertexElementUsage) keep 'int', because a cast to a smaller type would truncate an out of
  range value into a valid one. They are excluded with performance-enum-size.EnumIgnoreList.
- No stream operator prints one of these enums without casting it to uint32_t first.

+++ DemoApps/Shared/AntiAliasing/include/Shared/AntiAliasing/SupportedSampleCountFlags.hpp
@@ -33,9 +33,11 @@
 
 #include <FslBase/BasicTypes.hpp>
 
+#include <cstdint>
+
 namespace Fsl
 {
-  enum class SupportedSampleCountFlags : uint32_t
+  enum class SupportedSampleCountFlags : uint8_t
   {
     NoFlags = 0,
     Samples1 = 1,
