+++ DemoFramework/FslGraphics3D/BasicRender/include/FslGraphics3D/BasicRender/Shader/PredefinedShaderType.hpp
@@ -33,9 +33,11 @@
 
 #include <FslBase/BasicTypes.hpp>
 
+#include <cstdint>
+
 namespace Fsl::Graphics3D
 {
-  enum class PredefinedShaderType : uint16_t
+  enum class PredefinedShaderType : uint8_t
   {
     PositionColorTextureVertex = 0,
     PositionColorTextureFragment = 1,
