+++ DemoFramework/FslSimpleUI/Base/UnitTest/include/FslSimpleUI/Base/UnitTest/WindowMethod.hpp
@@ -33,7 +33,7 @@
 
 #include <cstdint>
 
-enum class WindowMethod : uint32_t
+enum class WindowMethod : uint16_t
 {
   WinInit = 0x0001,
   WinMarkLayoutAsDirty = 0x0002,
