+++ DemoFramework/FslSimpleUI/Render/IMBatch/source/FslSimpleUI/Render/IMBatch/MeshTransparencyFlags.hpp
@@ -33,9 +33,11 @@
 
 #include <FslBase/BasicTypes.hpp>
 
+#include <cstdint>
+
 namespace Fsl::UI::RenderIMBatch
 {
-  enum class MeshTransparencyFlags : uint32_t
+  enum class MeshTransparencyFlags : uint8_t
   {
     NoFlags = 0x00,
     Opaque = 0x01,
